umra_va_xaj/api/v1/User/serializer.py
from rest_framework import serializers

from umravaxajapp.models import AdminUser


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = AdminUser
        fields = "__all__"
