umra_va_xaj/requirements.txt
asgiref==3.5.2
Django==3.2.13
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.4
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
whitenoise==6.2.0
